api/requirements.txt
fastapi
google-genai
uvicorn[standard]